add material ui and edit data layer

+++ data/src/main/java/com/scholar/data/source/local/model/MaterialLocal.kt
@@ -11,7 +11,7 @@ data class MaterialLocal(
     override val id: Int,
     override val title: String,
     override val description: String?,
-    override val price: Double?,
+    override val price: Int?,
     override val content: String?,
 ) : Material
