+++ domain/src/main/java/com/scholar/domain/model/MaterialNetwork.kt
@@ -6,7 +6,7 @@ data class MaterialNetwork(
     override val id: Int,
     override val title: String,
     override val description: String?,
-    override val price: Double?,
+    override val price: Int?,
     override val content: String?,
 ) : Material
 
@@ -15,6 +15,6 @@ interface Material {
     val id: Int
     val title: String
     val description: String?
-    val price: Double?
+    val price: Int?
     val content: String?
 }
